requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==4.2.2
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
entrypoints==0.4
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langsmith==0.4.31
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==2.6.0
numpy==2.3.3
openai==1.109.1
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
protobuf==3.20.3
pyarrow==21.0.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
semver==3.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
streamlit==1.14.1
tenacity==9.1.2
tiktoken==0.11.0
toml==0.10.2
toolz==1.0.0
tornado==6.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
